test-front/requirements.txt
streamlit
websockets
streamlit-audiorecorder

